Major refactor to split the CLI into its own tool
