+++ .github/workflows/release.yml
@@ -26,11 +26,11 @@ jobs:
 
       - name: Build binaries
         run: |
-          GOOS=linux GOARCH=amd64 go build -o school_menu_connector-linux-amd64
-          GOOS=linux GOARCH=arm64 go build -o school_menu_connector-linux-arm64
-          GOOS=darwin GOARCH=amd64 go build -o school_menu_connector-darwin-amd64
-          GOOS=darwin GOARCH=arm64 go build -o school_menu_connector-darwin-arm64
-          GOOS=windows GOARCH=amd64 go build -o school_menu_connector-windows-amd64.exe
+          GOOS=linux GOARCH=amd64 go build -o school_menu_connector-linux-amd64 ./cmd/cli
+          GOOS=linux GOARCH=arm64 go build -o school_menu_connector-linux-arm64 ./cmd/cli
+          GOOS=darwin GOARCH=amd64 go build -o school_menu_connector-darwin-amd64 ./cmd/cli
+          GOOS=darwin GOARCH=arm64 go build -o school_menu_connector-darwin-arm64 ./cmd/cli
+          GOOS=windows GOARCH=amd64 go build -o school_menu_connector-windows-amd64.exe ./cmd/cli
 
       - name: Create Release
         id: create_release
